requirements.txt
Flask==1.1.2
gunicorn==20.1.0
Jinja2==2.11.3
Werkzeug==1.0.1
# numpy==1.22.1
matplotlib==3.3.4
MarkupSafe==1.1.1
plotly==5.6.0
scipy==1.7.2
xgboost==1.5.2
beautifulsoup4==4.10.0
torch==1.11.0
torchsummary==1.5.1
torchvision==0.12.0
urllib3==1.26.2
scikit-learn==0.24.2
statsmodels==0.13.0
lazy-object-proxy==1.4.3
Pillow==8.4.0
Flask-PyMongo==2.3.0
ImageHash==4.2.1
Cython==0.29.27
pandas==1.2.4
